requirements.txt
click==8.1.8
crewai==0.140.0
fastapi==0.110.3
uvicorn==0.29.0
python-multipart==0.0.9
starlette==0.37.2
openai==1.68.2
pydantic==2.9.2
pydantic_core==2.23.4
langchain-core==0.3.21
langchain-openai==0.2.9
langchain-community==0.3.9
langsmith==0.1.147
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.71.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
protobuf==5.28.3
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
numpy==2.1.0
pandas==2.2.2
onnxruntime==1.22.0
pillow==10.3.0
Jinja2==3.1.4
jsonschema==4.22.0
oauthlib==3.2.2
python-dotenv==1.0.0
pypdf==5.0.0
tiktoken==0.8.0